+++ noodles-gff/src/async.rs
@@ -1,0 +1,5 @@
+//! Async GFF.
+
+mod reader;
+
+pub use self::reader::Reader;